requirements.txt
flask==2.0.1
werkzeug==2.0.1
pytesseract==0.3.8
pdf2image==1.16.0
opencv-python==4.5.3.56
numpy==1.21.2
Pillow==8.3.2
pandas==1.3.3
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.1.3
sqlalchemy==1.4.23
price-parser==0.3.4
python-dateutil==2.8.2
logging==0.5.1.2
pathlib==1.0.1
typing==3.7.4.3
